script/totals works again!
Older years need some help (for example, if this year is not a leap year
and a timesheet has 2/29 as a date, we need to tell script/process what
year it should use).

Older timesheets may have historical employees instead of current ones.

Don't create a bunch of ${year}.txt in the root dir, it's ugly.